Correctly handle non-ASCII digits in radixes <= 10
There was a case in DIGIT-CHAR-P that was meant to deal with non-ASCII
digits, but it was shadowed by the fast path for base 10 and lower.
The fast path now has a guard on it that prevents it from engaging
when the char-code of the character that is being checked is above
ASCII.

Some more PARSE-INTEGER (which uses DIGIT-CHAR-P) test cases were
added to check the behavior of Unicode and other radixes.