Removes the mapping between BELL and #\Bel to fix Unicode
Unicode encodes U+1F514 (named BELL), which conflicts with the alias
"Bell" -> #\Bel defined in target-char.lisp. This causes
(name-char (char-name (code-char #x1F514))) => #\Bel not #\BELL
(U+1F514) as expected. This ambiguity had to be resolved by either
removing "BELL" -> #\Bel or "BELL" -> U+1F514. Discussion in IRC led
to the determination that removal of the #\Bel alias was the best
course of action.

This is a breaking change. If it affects you, please raise the issue
for further discussion.